backend/requirements.txt
```
flask==2.2.5
werkzeug==2.2.3
flask-cors==3.0.10
python-dotenv==0.21.0
SpeechRecognition==3.8.1
pyttsx3==2.90
pyautogui==0.9.53
openai==0.27.0
pyobjc-framework-Cocoa; sys_platform == 'darwin'
pyobjc-framework-QuartzCore; sys_platform == 'darwin'
pywin32; sys_platform == 'win32'
psutil==5.9.4
screenbrightness-win==0.1.0
screen-brightness-control==0.12.0
pyaudio==0.2.14
```